+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/PlanTargetContainer.cs
@@ -54,7 +54,7 @@ namespace Assistant.NINAPlugin.Sequencer {
         private readonly IPlanTarget previousPlanTarget;
         private readonly SchedulerPlan plan;
         private readonly IProfile activeProfile;
-        private SchedulerStatusMonitor monitor;
+        private SchedulerProgressVM schedulerProgress;
 
         private IImageSaveWatcher ImageSaveWatcher;
 
@@ -76,7 +76,7 @@ namespace Assistant.NINAPlugin.Sequencer {
                 IWindowServiceFactory windowServiceFactory,
                 IPlanTarget previousPlanTarget,
                 SchedulerPlan plan,
-                SchedulerStatusMonitor monitor) : base(new PlanTargetContainerStrategy()) {
+                SchedulerProgressVM schedulerProgress) : base(new PlanTargetContainerStrategy()) {
             Name = nameof(PlanTargetContainer);
             Description = "";
             Category = "Assistant";
@@ -97,13 +97,13 @@ namespace Assistant.NINAPlugin.Sequencer {
             this.plateSolverFactory = plateSolverFactory;
             this.windowServiceFactory = windowServiceFactory;
 
-            this.monitor = monitor;
+            this.schedulerProgress = schedulerProgress;
             this.previousPlanTarget = previousPlanTarget;
             this.plan = plan;
             this.activeProfile = profileService.ActiveProfile;
 
             PlanTargetContainerStrategy containerStrategy = Strategy as PlanTargetContainerStrategy;
-            containerStrategy.SetContext(parentContainer, plan, monitor);
+            containerStrategy.SetContext(parentContainer, plan, schedulerProgress);
             AttachNewParent(parentContainer);
 
             if (!plan.IsEmulator)
@@ -187,7 +187,6 @@ namespace Assistant.NINAPlugin.Sequencer {
 
                 if (instruction is PlanSlew) {
                     AddSlew((PlanSlew)instruction, plan.PlanTarget);
-                    //Notification.ShowInformation("REMINDER: SKIPPING SLEW");
                     continue;
                 }
 
@@ -216,6 +215,11 @@ namespace Assistant.NINAPlugin.Sequencer {
                     continue;
                 }
 
+                if (instruction is PlanBeforeTargetContainer) {
+                    AddBeforeTargetInstructions();
+                    continue;
+                }
+
                 TSLogger.Error($"unknown instruction type: {instruction.GetType().FullName}");
                 throw new Exception($"unknown instruction type: {instruction.GetType().FullName}");
             }
@@ -226,9 +230,6 @@ namespace Assistant.NINAPlugin.Sequencer {
             bool isPlateSolve = instruction.center;
             InputCoordinates slewCoordinates = new InputCoordinates(planTarget.Coordinates);
             SequenceItem slewCenter;
-
-            //isPlateSolve = false;
-            //Notification.ShowInformation("REMINDER: center is disabled for slews");
 
             string with = isPlateSolve ? "with" : "without";
             TSLogger.Info($"slew ({with} center): {Utils.FormatCoordinates(planTarget.Coordinates)}");
@@ -254,6 +255,13 @@ namespace Assistant.NINAPlugin.Sequencer {
 
             SetItemDefaults(slewCenter, null);
             Add(slewCenter);
+        }
+
+        private void AddBeforeTargetInstructions() {
+            if (parentContainer.BeforeTargetContainer.Items?.Count > 0) {
+                parentContainer.BeforeTargetContainer.ResetAll();
+                Add(parentContainer.BeforeTargetContainer);
+            }
         }
 
         private void AddSwitchFilter(IPlanExposure planExposure) {